add vllm_collector and test_vllm
