+++ ding/worker/collector/vllm_collector.py
@@ -3,10 +3,10 @@ import os
 import uuid
 from loguru import logger
 from vllm import AsyncLLMEngine, AsyncEngineArgs, SamplingParams, RequestOutput
-from vllm.assets.image import ImageAsset
+
 
 class VllmActor:
-    def __init__(self, model_path: str) -> None:
+    def __init__(self, model_path: str,mm_processor_kwargs: dict) -> None:
         """
         Overview:
             Initialize the vLLM actor. For more details, please refer to https://docs.vllm.ai/en/stable.
@@ -19,6 +19,7 @@ class VllmActor:
         # Set CUDA_VISIBLE_DEVICES to use only free GPUs
         os.environ["CUDA_VISIBLE_DEVICES"] = ",".join(map(str, self.free_gpus))
         self.model_path = model_path
+        self.mm_processor_kwargs=mm_processor_kwargs
         self._initialize()
 
     def get_free_gpus(self) -> List[int]:
@@ -56,17 +57,10 @@ class VllmActor:
             tensor_parallel_size=self.num_gpus,
             max_num_batched_tokens=8192,
             max_model_len=8192,
-            #max_model_len=4096,  #see if 8192 works
-            #max_num_batched_tokens=4096,
-            #max_num_batched_tokens=2048,
-            #max_model_len=2048,
             # enable_chunked_prefill=True,
             max_num_seqs=5,
             # Note - mm_processor_kwargs can also be passed to generate/chat calls
-            mm_processor_kwargs={
-                "min_pixels": 28 * 28,
-                "max_pixels": 1280 * 28 * 28,
-            },
+            mm_processor_kwargs=self.mm_processor_kwargs,
         )
         self.engine = AsyncLLMEngine.from_engine_args(engine_args)
 
@@ -106,7 +100,10 @@ class HuggingFaceModelGenerator:
         A LLM/VLM generator that uses Hugging Face models with vLLM as the backend.
     """
 
-    def __init__(self, model_path: str, max_tokens: int = 1024, temperature: float = 0) -> None:
+    def __init__(self, model_path: str, max_tokens: int = 1024, temperature: float = 0, mm_processor_kwargs:dict = {
+                "min_pixels": 28 * 28,
+                "max_pixels": 1280 * 28 * 28,
+            }) -> None:
         """
         Overview:
             Initialize the Hugging Face model generator.
@@ -115,7 +112,7 @@ class HuggingFaceModelGenerator:
             - max_tokens (int): The maximum number of tokens to generate, default to 1024.
             - temperature (float): The temperature for the language model, default to 0.
         """
-        self.vllm_actor = VllmActor(model_path)
+        self.vllm_actor = VllmActor(model_path,mm_processor_kwargs)
         self.max_tokens = max_tokens
         self.temperature = temperature
 
@@ -144,79 +141,3 @@ class HuggingFaceModelGenerator:
             for x, conf in zip(response.outputs, confidence_scores)
         ]
         
-
-model=HuggingFaceModelGenerator('/mnt/afs/share/Qwen2-VL-7B',temperature=0.5) #设置一个temperature就好了,可以做到生成多个候选答案
-
-def get_prompts_qwen(questions: list,modality: str):
-    if modality == "image":
-        placeholder = "<|image_pad|>"
-    elif modality == "video":
-        placeholder = "<|video_pad|>"
-
-    prompts = [("<|im_start|>system\nYou are a helpful assistant.<|im_end|>\n"
-              f"<|im_start|>user\n<|vision_start|>{placeholder}<|vision_end|>"
-              f"{question}<|im_end|>\n"
-              "<|im_start|>assistant\n") for question in questions]
-    stop_token_ids = None
-    return prompts,stop_token_ids
-
-def get_multi_modal_input(modality,filenames,questions):
-    """
-    return {
-        "data": image or video,
-        "question": question,
-    }
-    """
-    if modality == "image":
-        # Input image and question
-        ret={}
-        ret["data"]=[]
-        ret["question"]=[]
-        for filename,question in zip(filenames,questions):
-            image = ImageAsset(filename) \
-                .pil_image.convert("RGB")
-            #img_question = "What is the content of this image?"
-            img_question=question
-            ret["data"].append(image)
-            ret["question"].append(img_question)
-        return ret
-
-
-questions=["What is the content of this image?","Please describe the image.","How many people are there in the image? What are they doing?"]
-img_names=['/mnt/afs/niuyazhe/data/meme/data/Eimages/Eimages/Eimages/image_ (2)','/mnt/afs/niuyazhe/data/meme/data/Cimages/Cimages/Cimages/Image_(3127)','/mnt/afs/wangqijian/data/test/test']
-#questions=["What is the content of this image?"]
-#img_names=['/mnt/afs/niuyazhe/data/meme/data/Eimages/Eimages/Eimages/image_ (2)']
-num_prompts=len(questions)
-image_repeat_prob=None
-modality = 'image'
-
-mm_input = get_multi_modal_input(modality,img_names,questions)
-data = mm_input["data"]
-question = mm_input["question"]
-batch_inference_mine=True
-prompts, stop_token_ids = get_prompts_qwen(question,modality)
-
-
-import asyncio
-import nest_asyncio
-nest_asyncio.apply()
-async def main():
-    inputs = [
-        {
-            "prompt":prompt,
-            "multi_modal_data":{
-                modality:data
-            }
-        } for prompt,data in zip(prompts,data)
-    ]
-    # 调用 generate 方法
-    for in_data in inputs:
-        tactics = await model.generate(prompt=in_data, num_samples=3)
-        # 打印返回结果
-        for tactic, confidence in tactics:
-            print(f"Tactic: {tactic}")
-    
-
-# 运行主程序
-if __name__ == "__main__":
-    asyncio.run(main())
